fix(jz): 补齐 Memfit 侧 yaklang 源码修复与跨平台构建脚本
对 JZ_Memfit 仓库 master(83f32a5bf 及之后)做全量对齐后的补漏:

- vpnbrute/pptp.go: 0xffffffff 加 uint 转换, 修复 32 位平台编译
- httptpl/nuclei_dsl.go: math.MaxInt64 改 math.MaxInt, 修复 32 位 rand.Intn 溢出
- loop_fund_analysis reflection 示例更新为 excelcli 工作流(取 Memfit 最新版;
  其余 4 个 loop 的 prompts 官方 7 月版更新, 保留不动)
- 新增 build_yak_linux_amd64.sh / build_yak_windows.sh / build_yak_windows_x86.sh
  交叉编译脚本(与 JZ_0512、JZ_Memfit 两处版本核对一致)

+++ common/vpnbrute/pptp.go
@@ -225,8 +225,8 @@ func (pptp *PPTPAuthItem) GetSetLinkInfo() map[string]any {
 		"Set Link Info": map[string]any{
 			"PeerCallId": pptp.cfg.PeerCallId,
 			"Reserved":   0,
-			"Send Accm":  0xffffffff,
-			"Recv Accm":  0xffffffff,
+			"Send Accm":  uint(0xffffffff),
+			"Recv Accm":  uint(0xffffffff),
 		},
 	}
 }
